Listen on All Hosts in Dockerfile; Add README
After having tested the project directly from GitHub, I could run with Flask locally and it worked fine. I could also build and deploy a Docker container and the Docker container reported no issues serving the Flask application. However, the Docker container was not exposing access to all hosts, only all ports. For whatever reason the "host=0.0.0.0" line in the "app.py" file was having no impact on access. The fix to expose host access was in the runtime of the application which is reflected in the Dockerfile. After adding this I could access everything normally from my host to the container.

Also added a bit of README documentation for building and running with Docker which will help with testing. Can be removed later if you want.